Refactor reservation units to own file, add more fields to serializers
